+++ ui/__init__.py
@@ -1,0 +1,1 @@
+# Make UI package importable
