+++ ui/board.py
@@ -1,859 +1,17 @@
 from PyQt5.QtWidgets import (
-    QApplication, QMainWindow, QGridLayout, QLabel, QVBoxLayout, 
-    QHBoxLayout, QComboBox, QPushButton, QDialog, QWidget, QSlider,
-    QListWidget, QSplitter, QFrame, QGraphicsOpacityEffect, QListWidgetItem,
-    QScrollArea
+    QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QGridLayout, QLabel, 
+    QSplitter, QFrame
 )
-from PyQt5.QtCore import Qt, QTimer, QPropertyAnimation, QPoint, QEasingCurve, pyqtSignal, QSize, QThread, QRect
-from PyQt5.QtGui import QColor, QIcon, QFont, QPalette, QBrush, QFontDatabase
+from PyQt5.QtCore import Qt, QPoint, QTimer
 import chess
 import sys
-import time
-import os
-from chess_engine import find_best_move as find_best_move1
-from chess_engine2 import find_best_move as find_best_move2
 
-# Thread class for AI computation to prevent GUI freezing
-class AIWorker(QThread):
-    finished = pyqtSignal(str)
-    
-    def __init__(self, fen, depth, engine_num=1):
-        super().__init__()
-        self.fen = fen
-        self.depth = depth
-        self.engine_num = engine_num
-        
-    def run(self):
-        try:
-            if self.engine_num == 1:
-                result = find_best_move1(self.fen, self.depth)
-            else:
-                result = find_best_move2(self.fen, self.depth)
-            self.finished.emit(result)
-        except Exception as e:
-            print(f"AI Worker error: {str(e)}")
-            self.finished.emit("")
-
-class AnimatedLabel(QLabel):
-    """Custom QLabel with animation capabilities for chess pieces"""
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.animation = QPropertyAnimation(self, b"pos")
-        self.animation.setEasingCurve(QEasingCurve.OutCubic)
-        self.animation.setDuration(300)  # 300ms for piece movement
-        
-    def move_to(self, target_pos):
-        """Animate movement to target position"""
-        start_pos = self.pos()
-        self.animation.setStartValue(start_pos)
-        self.animation.setEndValue(target_pos)
-        self.animation.start()
-
-class ChessSquare(QLabel):
-    """Enhanced chess square widget with hover and selection effects"""
-    clicked = pyqtSignal(int, int)
-    
-    def __init__(self, row, col, parent=None):
-        super().__init__(parent)
-        self.row = row
-        self.col = col
-        self.setAlignment(Qt.AlignCenter)
-        self.setFixedSize(60, 60)
-        self.setMargin(0)
-        
-        # Initialize states
-        self.is_highlighted = False
-        self.is_last_move = False
-        self.is_valid_move = False
-        self.is_castling_move = False
-        self.is_selected = False
-        
-    def enterEvent(self, event):
-        """Highlight square on mouse hover"""
-        if not self.is_selected and not self.is_last_move:
-            effect = QGraphicsOpacityEffect(self)
-            effect.setOpacity(0.8)
-            self.setGraphicsEffect(effect)
-        super().enterEvent(event)
-        
-    def leaveEvent(self, event):
-        """Remove highlight on mouse leave"""
-        if not self.is_selected and not self.is_last_move:
-            self.setGraphicsEffect(None)
-        super().leaveEvent(event)
-    
-    def mousePressEvent(self, event):
-        """Handle mouse click on square"""
-        self.clicked.emit(self.row, self.col)
-        super().mousePressEvent(event)
-        
-    def update_appearance(self):
-        """Update the square's appearance based on its state"""
-        base_color = "#c1bfb0" if (self.row + self.col) % 2 == 0 else "#7a9bbe"
-        
-        if self.is_selected:
-            base_color = "#ffff99"  # Yellow for selected piece
-        elif self.is_last_move:
-            base_color = "#ffe066"  # Soft yellow for last move
-        elif self.is_castling_move:
-            base_color = "#ff9966"  # Orange for castling moves
-        elif self.is_valid_move:
-            base_color = "#90ee90"  # Light green for valid moves
-            
-        self.setStyleSheet(f"background-color: {base_color}; border: 1px solid black;")
-
-class StartScreen(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Chess Game - Mode Selection")
-        self.setModal(True)
-        self.setFixedSize(500, 400)
-        self.setStyleSheet("""
-            QDialog {
-                background-color: #f0f0f0;
-                border-radius: 10px;
-            }
-            QLabel {
-                color: #333333;
-            }
-        """)
-        
-        layout = QVBoxLayout()
-        
-        title_label = QLabel("Chess Game")
-        title_label.setAlignment(Qt.AlignCenter)
-        title_label.setStyleSheet("font-size: 36pt; font-weight: bold; margin-bottom: 30px; color: #2c3e50;")
-        layout.addWidget(title_label)
-        
-        subtitle = QLabel("Select Game Mode:")
-        subtitle.setAlignment(Qt.AlignCenter)
-        subtitle.setStyleSheet("font-size: 18pt; margin-bottom: 30px;")
-        layout.addWidget(subtitle)
-        
-        # Button container with attractive styling
-        button_container = QWidget()
-        button_container.setStyleSheet("""
-            QWidget {
-                background-color: transparent;
-                border-radius: 15px;
-            }
-        """)
-        buttons_layout = QVBoxLayout(button_container)
-        
-        self.human_ai_button = QPushButton("Human vs AI")
-        self.human_ai_button.setFixedHeight(70)
-        self.human_ai_button.setStyleSheet("""
-            QPushButton {
-                background-color: #3498db;
-                color: white;
-                font-size: 18pt;
-                padding: 15px;
-                border-radius: 10px;
-                margin-bottom: 20px;
-            }
-            QPushButton:hover {
-                background-color: #2980b9;
-            }
-            QPushButton:pressed {
-                background-color: #1c6ea4;
-            }
-        """)
-        
-        self.ai_ai_button = QPushButton("AI vs AI")
-        self.ai_ai_button.setFixedHeight(70)
-        self.ai_ai_button.setStyleSheet("""
-            QPushButton {
-                background-color: #e74c3c;
-                color: white;
-                font-size: 18pt;
-                padding: 15px;
-                border-radius: 10px;
-            }
-            QPushButton:hover {
-                background-color: #c0392b;
-            }
-            QPushButton:pressed {
-                background-color: #a93226;
-            }
-        """)
-        
-        buttons_layout.addWidget(self.human_ai_button)
-        buttons_layout.addWidget(self.ai_ai_button)
-        layout.addWidget(button_container)
-        
-        self.human_ai_button.clicked.connect(self.choose_human_ai)
-        self.ai_ai_button.clicked.connect(self.choose_ai_ai)
-        
-        self.game_mode = None
-        self.setLayout(layout)
-    
-    def choose_human_ai(self):
-        self.game_mode = "human_ai"
-        self.accept()
-    
-    def choose_ai_ai(self):
-        self.game_mode = "ai_ai"
-        self.accept()
-    
-    def get_mode(self):
-        return self.game_mode
-
-class PawnPromotionDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Pawn Promotion")
-        self.setModal(True)
-        self.setFixedSize(300, 250)
-        self.setStyleSheet("""
-            QDialog {
-                background-color: #f5f5f5;
-                border-radius: 10px;
-            }
-        """)
-
-        layout = QVBoxLayout()
-
-        self.label = QLabel("Choose a piece to promote to:")
-        self.label.setAlignment(Qt.AlignCenter)
-        self.label.setStyleSheet("font-size: 16pt; margin-bottom: 20px; color: #333;")
-        layout.addWidget(self.label)
-
-        self.promotion_choice = QComboBox()
-        self.promotion_choice.setStyleSheet("""
-            QComboBox {
-                font-size: 14pt;
-                padding: 8px;
-                border: 2px solid #bdc3c7;
-                border-radius: 6px;
-                background-color: white;
-                color: #333;
-            }
-            QComboBox::drop-down {
-                border: 0px;
-                width: 30px;
-            }
-            QComboBox QAbstractItemView {
-                background-color: white;
-                selection-background-color: #3498db;
-                selection-color: white;
-                color: #333;
-            }
-        """)
-        self.promotion_choice.addItems(["Queen", "Rook", "Bishop", "Knight"])
-        layout.addWidget(self.promotion_choice)
-
-        self.ok_button = QPushButton("Promote")
-        self.ok_button.setStyleSheet("""
-            QPushButton {
-                background-color: #2ecc71;
-                color: white;
-                font-size: 14pt;
-                font-weight: bold;
-                padding: 10px;
-                border-radius: 6px;
-                margin-top: 20px;
-            }
-            QPushButton:hover {
-                background-color: #27ae60;
-            }
-            QPushButton:pressed {
-                background-color: #219653;
-            }
-        """)
-        self.ok_button.clicked.connect(self.accept)
-        layout.addWidget(self.ok_button)
-
-        self.setLayout(layout)
-
-    def get_choice(self):
-        piece_map = {
-            "Queen": "q",
-            "Rook": "r",
-            "Bishop": "b",
-            "Knight": "n"
-        }
-        return piece_map[self.promotion_choice.currentText()]
-
-class GameOverPopup(QDialog):
-    play_again_signal = pyqtSignal()
-    return_home_signal = pyqtSignal()
-    
-    def __init__(self, result, parent=None, custom_message=None):
-        super().__init__(parent)
-        self.setWindowTitle("Game Over")
-        self.setWindowFlags(Qt.Dialog | Qt.FramelessWindowHint)  
-        self.setModal(True)
-        self.setFixedSize(450, 300)  # Made taller for additional buttons
-        
-        self.setStyleSheet("""
-            QDialog {
-                background-color: white;
-                border: 3px solid #4CAF50;
-                border-radius: 15px;
-            }
-        """)
-        
-        layout = QVBoxLayout()
-        layout.setContentsMargins(25, 25, 25, 25)
-
-        if custom_message:
-            message = custom_message
-            if "AI 1" in message:
-                color = "#4CAF50"
-                self.setStyleSheet("QDialog { background-color: white; border: 3px solid #4CAF50; border-radius: 15px; }")
-            elif "AI 2" in message:
-                color = "#F44336"
-                self.setStyleSheet("QDialog { background-color: white; border: 3px solid #F44336; border-radius: 15px; }")
-            else:
-                color = "#2196F3"
-                self.setStyleSheet("QDialog { background-color: white; border: 3px solid #2196F3; border-radius: 15px; }")
-        else:
-            if result == '1-0':
-                message = "🏆 Player (White) Wins! 🏆"
-                color = "#4CAF50"
-            elif result == '0-1':
-                message = "❌ AI (Black) Wins! ❌"
-                color = "#F44336"
-            else:
-                message = "🤝 It's a Draw! 🤝"
-                color = "#2196F3"
-
-        # Add a decorative header
-        header = QLabel("GAME OVER", self)
-        header.setAlignment(Qt.AlignCenter)
-        header.setStyleSheet(f"""
-            font-size: 24pt; 
-            font-weight: bold; 
-            color: {color}; 
-            padding: 10px;
-            font-family: 'Arial', sans-serif;
-            border-bottom: 2px solid {color};
-        """)
-        layout.addWidget(header)
-
-        label = QLabel(message, self)
-        label.setAlignment(Qt.AlignCenter)
-        label.setStyleSheet(f"""
-            font-size: 22pt; 
-            font-weight: bold; 
-            color: {color}; 
-            padding: 20px;
-            font-family: 'Arial', sans-serif;
-        """)
-        layout.addWidget(label)
-
-        # Button container
-        button_layout = QHBoxLayout()
-        
-        # Play Again button
-        self.play_again_button = QPushButton("Play Again", self)
-        self.play_again_button.setCursor(Qt.PointingHandCursor)
-        self.play_again_button.setStyleSheet(f"""
-            QPushButton {{
-                background-color: {color};
-                color: white;
-                font-size: 14pt;
-                font-weight: bold;
-                padding: 12px 20px;
-                border-radius: 8px;
-                border: none;
-                min-width: 150px;
-            }}
-            QPushButton:hover {{
-                background-color: #{color[1:]}dd;
-            }}
-            QPushButton:pressed {{
-                background-color: #{color[1:]}aa;
-            }}
-        """)
-        self.play_again_button.clicked.connect(self.play_again)
-        
-        # Return Home button
-        self.return_home_button = QPushButton("Return to Home", self)
-        self.return_home_button.setCursor(Qt.PointingHandCursor)
-        self.return_home_button.setStyleSheet(f"""
-            QPushButton {{
-                background-color: #607D8B;
-                color: white;
-                font-size: 14pt;
-                font-weight: bold;
-                padding: 12px 20px;
-                border-radius: 8px;
-                border: none;
-                min-width: 180px;
-            }}
-            QPushButton:hover {{
-                background-color: #78909C;
-            }}
-            QPushButton:pressed {{
-                background-color: #455A64;
-            }}
-        """)
-        self.return_home_button.clicked.connect(self.return_home)
-        
-        button_layout.addWidget(self.play_again_button)
-        button_layout.addWidget(self.return_home_button)
-        
-        layout.addLayout(button_layout)
-        layout.addSpacing(10)
-
-        self.setLayout(layout)
-    
-    def play_again(self):
-        self.play_again_signal.emit()
-        self.accept()
-    
-    def return_home(self):
-        self.return_home_signal.emit()
-        self.accept()
-
-class MoveHistoryWidget(QFrame):
-    """Widget to display the move history with improved contrast"""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setFrameShape(QFrame.StyledPanel)
-        self.setStyleSheet("""
-            QFrame {
-                background-color: #ffffff;
-                border-radius: 8px;
-                border: 1px solid #cccccc;
-            }
-        """)
-        
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(10, 10, 10, 10)
-        layout.setSpacing(8)
-        
-        # Title container with improved visibility
-        title_container = QFrame()
-        title_container.setStyleSheet("""
-            QFrame {
-                background-color: #34495e;
-                border-radius: 6px;
-                border: 1px solid #1a2530;
-            }
-        """)
-        
-        title_layout = QVBoxLayout(title_container)
-        
-        # Title label
-        title = QLabel("Move History")
-        title.setAlignment(Qt.AlignCenter)
-        title.setStyleSheet("""
-            font-size: 16pt; 
-            font-weight: bold; 
-            color: white; 
-            padding: 5px;
-        """)
-        title_layout.addWidget(title)
-        layout.addWidget(title_container)
-        
-        # Custom styled list widget
-        self.move_list = QListWidget()
-        self.move_list.setStyleSheet("""
-            QListWidget {
-                background-color: #ffffff;
-                alternate-background-color: #f0f0f0;
-                font-size: 14pt;
-                font-family: 'Arial', monospace;
-                border: 1px solid #cccccc;
-                border-radius: 4px;
-                padding: 5px;
-            }
-            QListWidget::item {
-                color: #222222;
-                padding: 5px;
-                border-bottom: 1px solid #e0e0e0;
-                min-height: 30px;
-            }
-            QListWidget::item:selected {
-                background-color: #3498db;
-                color: white;
-            }
-            QListWidget::item:hover {
-                background-color: #e6f7ff;
-            }
-            QListWidget::item:alternate {
-                background-color: #f5f5f5;
-            }
-        """)
-        self.move_list.setAlternatingRowColors(True)
-        layout.addWidget(self.move_list)
-        
-    def add_move(self, piece, from_square, to_square, color="White", capture=False, check=False, promotion=None, castling=False):
-        """Add a move to the history with proper chess notation"""
-        piece_symbols = {
-            chess.PAWN: "",
-            chess.KNIGHT: "N",
-            chess.BISHOP: "B",
-            chess.ROOK: "R",
-            chess.QUEEN: "Q",
-            chess.KING: "K"
-        }
-        
-        notation = ""
-        
-        # Add move number for White moves
-        if color == "White":
-            move_number = (self.move_list.count() // 2) + 1
-            notation = f"{move_number}. "
-        
-        # Handle castling notation
-        if castling:
-            if to_square[0] > from_square[0]:  # King-side castling
-                notation += "O-O"
-            else:  # Queen-side castling
-                notation += "O-O-O"
-        else:
-            # Add piece symbol
-            notation += piece_symbols.get(piece.piece_type, "")
-            
-            # Add capture symbol
-            if capture:
-                if piece.piece_type == chess.PAWN:
-                    notation += from_square[0] + "x"
-                else:
-                    notation += "x"
-                    
-            # Add destination square
-            notation += to_square
-            
-            # Add promotion piece
-            if promotion:
-                notation += "=" + piece_symbols.get(promotion, "")
-                
-        # Add check/checkmate symbol
-        if check:
-            notation += "+"
-            
-        # Format the move in the list with improved styling
-        if color == "White":
-            item = QListWidgetItem(f"{notation.ljust(12)}")
-            item_color = "#0000aa"  # Dark blue for white's moves
-            item.setForeground(QBrush(QColor(item_color)))
-            self.move_list.addItem(item)
-        else:
-            # Update the last item to include the black move
-            current_item = self.move_list.item(self.move_list.count() - 1)
-            if current_item:
-                current_text = current_item.text()
-                combined_text = f"{current_text} {notation}"
-                current_item.setText(combined_text)
-                
-                # Apply custom styling with a bold font
-                font = current_item.font()
-                font.setBold(True)
-                current_item.setFont(font)
-            
-        # Scroll to the bottom
-        self.move_list.scrollToBottom()
-
-    def clear_history(self):
-        """Clear the move history"""
-        self.move_list.clear()
-
-class ThinkingIndicator(QLabel):
-    """Visual indicator for AI thinking state with improved visibility"""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setAlignment(Qt.AlignCenter)
-        self.setStyleSheet("""
-            font-size: 16pt;
-            font-weight: bold;
-            color: white;
-            background-color: rgba(52, 73, 94, 0.8);
-            border-radius: 10px;
-            padding: 10px;
-            border: 1px solid white;
-            margin: 0px;
-        """)
-        self.setFixedHeight(50)
-        self.dots = 0
-        self.timer = QTimer(self)
-        self.timer.timeout.connect(self.update_dots)
-        self.hide()
-        
-    def start_thinking(self, ai_name):
-        """Start the thinking animation"""
-        self.base_text = f"{ai_name} thinking"
-        self.dots = 0
-        self.setText(f"{self.base_text}...")
-        self.show()
-        self.timer.start(300)  # Update dots every 300ms
-        
-    def stop_thinking(self):
-        """Stop the thinking animation"""
-        self.timer.stop()
-        self.hide()
-        
-    def update_dots(self):
-        """Update the thinking dots animation"""
-        self.dots = (self.dots + 1) % 4
-        dot_text = "." * self.dots
-        self.setText(f"{self.base_text}{dot_text.ljust(3)}")
-
-class ControlButton(QPushButton):
-    """Enhanced button with better visual feedback"""
-    def __init__(self, text, color, icon=None, parent=None):
-        super().__init__(text, parent)
-        self.base_color = color
-        self.setCursor(Qt.PointingHandCursor)
-        self.setFixedHeight(50)  # Adjusted for better appearance on small screens
-        
-        if icon:
-            self.setIcon(icon)
-            self.setIconSize(QSize(24, 24))
-        
-        self.updateStyle()
-    
-    def updateStyle(self):
-        self.setStyleSheet(f"""
-            QPushButton {{
-                background-color: {self.base_color};
-                color: white;
-                font-size: 14pt;
-                font-weight: bold;
-                padding: 8px 16px;
-                border-radius: 8px;
-                border: none;
-                text-align: center;
-            }}
-            QPushButton:hover {{
-                background-color: {self._lighten_color(self.base_color, 1.1)};
-            }}
-            QPushButton:pressed {{
-                background-color: {self._darken_color(self.base_color, 1.1)};
-            }}
-            QPushButton:disabled {{
-                background-color: #bbbbbb;
-                color: #dddddd;
-            }}
-        """)
-    
-    def _lighten_color(self, color, factor):
-        # Simple implementation to lighten a hex color
-        if color.startswith('#'):
-            color = color[1:]
-        r = min(255, int(int(color[0:2], 16) * factor))
-        g = min(255, int(int(color[2:4], 16) * factor))
-        b = min(255, int(int(color[4:6], 16) * factor))
-        return f"#{r:02x}{g:02x}{b:02x}"
-    
-    def _darken_color(self, color, factor):
-        # Simple implementation to darken a hex color
-        if color.startswith('#'):
-            color = color[1:]
-        r = max(0, int(int(color[0:2], 16) / factor))
-        g = max(0, int(int(color[2:4], 16) / factor))
-        b = max(0, int(int(color[4:6], 16) / factor))
-        return f"#{r:02x}{g:02x}{b:02x}"
-
-# Modify the slider label colors for better visibility
-class EnhancedSlider(QWidget):
-    """Custom slider with better visual design and labels"""
-    valueChanged = pyqtSignal(int)
-    
-    def __init__(self, title, min_val, max_val, default_val, min_label, max_label, parent=None):
-        super().__init__(parent)
-        
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Title with improved visibility
-        title_container = QFrame()
-        title_container.setStyleSheet("""
-            QFrame {
-                background-color: #e8e8e8;
-                border-radius: 6px;
-                border: 1px solid #cccccc;
-            }
-        """)
-        title_layout = QVBoxLayout(title_container)
-        title_layout.setContentsMargins(10, 8, 10, 8)
-        
-        title_label = QLabel(title)
-        title_label.setAlignment(Qt.AlignCenter)
-        title_label.setStyleSheet("""
-            font-size: 12pt; 
-            font-weight: bold; 
-            color: #333333;
-        """)
-        title_layout.addWidget(title_label)
-        layout.addWidget(title_container)
-        
-        # Create the styled slider
-        self.slider = QSlider(Qt.Horizontal)
-        self.slider.setMinimum(min_val)
-        self.slider.setMaximum(max_val)
-        self.slider.setValue(default_val)
-        self.slider.setTickPosition(QSlider.TicksBelow)
-        self.slider.setTickInterval((max_val - min_val) // 4)
-        self.slider.setStyleSheet("""
-            QSlider {
-                height: 25px;
-                margin: 5px 0;
-            }
-            QSlider::groove:horizontal {
-                height: 8px;
-                background: #d0d0d0;
-                border-radius: 4px;
-            }
-            QSlider::handle:horizontal {
-                background: #2196F3;
-                width: 18px;
-                height: 18px;
-                margin: -5px 0;
-                border-radius: 9px;
-                border: 1px solid #1976D2;
-            }
-            QSlider::sub-page:horizontal {
-                background: #64B5F6;
-                border-radius: 4px;
-            }
-        """)
-        self.slider.valueChanged.connect(self._emit_value_changed)
-        layout.addWidget(self.slider)
-        
-        # Add min/max labels with IMPROVED CONTRAST
-        labels_layout = QHBoxLayout()
-        min_text = QLabel(min_label)
-        min_text.setStyleSheet("color: white; font-weight: bold; font-size: 10pt;")  # Changed to white for better contrast
-        
-        max_text = QLabel(max_label)
-        max_text.setAlignment(Qt.AlignRight)
-        max_text.setStyleSheet("color: white; font-weight: bold; font-size: 10pt;")  # Changed to white for better contrast
-        
-        labels_layout.addWidget(min_text)
-        labels_layout.addWidget(max_text)
-        layout.addLayout(labels_layout)
-        
-    def _emit_value_changed(self, value):
-        self.valueChanged.emit(value)
-        
-    def value(self):
-        return self.slider.value()
-    
-    def setValue(self, value):
-        self.slider.setValue(value)
-
-# Redesigned AI Control Panel with responsive layout
-class AIControlPanel(QScrollArea):
-    """AI control panel with improved responsive design"""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setFrameShape(QFrame.StyledPanel)
-        self.setWidgetResizable(True)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setStyleSheet("""
-            QScrollArea {
-                background-color: #2c3e50;  /* Darker blue background for better contrast */
-                border-radius: 10px;
-                border: 2px solid #1a2530;  /* Darker border */
-            }
-        """)
-        
-        # Create inner widget for the scroll area
-        inner_widget = QWidget()
-        inner_widget.setStyleSheet("background-color: #2c3e50;")  
-        self.setWidget(inner_widget)
-        
-        # Create main layout with adaptive spacing
-        main_layout = QVBoxLayout(inner_widget)
-        main_layout.setSpacing(10)
-        main_layout.setContentsMargins(10, 10, 10, 10)
-        
-        # Title header
-        title = QLabel("AI Controls")
-        title.setAlignment(Qt.AlignCenter)
-        title.setStyleSheet("""
-            font-size: 16pt; 
-            font-weight: bold; 
-            color: white; 
-            padding: 8px;
-            background-color: #34495e;  /* Slightly lighter than panel for contrast */
-            border-radius: 6px;
-        """)
-        main_layout.addWidget(title)
-                
-        # Create a responsive layout that can adapt to window width
-        responsive_container = QWidget()
-        responsive_layout = QHBoxLayout(responsive_container)
-        responsive_layout.setSpacing(10)
-        
-        # Create button container
-        button_container = QWidget()
-        button_layout = QVBoxLayout(button_container)
-        button_layout.setSpacing(8)
-        
-        # Start button - brighter green
-        self.start_button = ControlButton("▶ Start", "#2ecc71")  # Brighter green
-
-        # Pause button - brighter orange
-        self.pause_button = ControlButton("⏸ Pause", "#f39c12")  # Brighter orange
-
-        # Reset button - brighter blue
-        self.reset_button = ControlButton("↻ Reset", "#3498db")  # Brighter blue
-
-        # Return to Home button - brighter red
-        self.home_button = ControlButton("🏠 Home", "#e74c3c")  # Brighter red
-        
-        button_layout.addWidget(self.start_button)
-        button_layout.addWidget(self.pause_button)
-        button_layout.addWidget(self.reset_button)
-        button_layout.addWidget(self.home_button)
-        button_layout.addStretch(1)
-        
-        # Add button container to responsive layout
-        responsive_layout.addWidget(button_container, 1)
-        
-        # Create slider container
-        slider_container = QWidget()
-        slider_layout = QVBoxLayout(slider_container)
-        slider_layout.setSpacing(10)
-        
-        # Create enhanced sliders
-        self.speed_slider = EnhancedSlider(
-            "AI Move Speed:", 200, 2000, 800, "Fast", "Slow"
-        )
-        slider_layout.addWidget(self.speed_slider)
-        
-        self.depth_slider = EnhancedSlider(
-            "AI Thinking Depth:", 2, 5, 3, "Shallow", "Deep"
-        )
-        slider_layout.addWidget(self.depth_slider)
-        
-        # Add current depth display with improved visibility
-        depth_container = QFrame()
-        depth_container.setStyleSheet("""
-            QFrame {
-                background-color: #333333;
-                border-radius: 5px;
-                padding: 6px;
-            }
-        """)
-        
-        depth_layout = QVBoxLayout(depth_container)
-        self.depth_value = QLabel("Current depth: 3")
-        self.depth_value.setAlignment(Qt.AlignCenter)
-        self.depth_value.setStyleSheet("""
-            font-size: 12pt;
-            font-weight: bold;
-            color: white;
-            padding: 2px;
-        """)
-        depth_layout.addWidget(self.depth_value)
-        slider_layout.addWidget(depth_container)
-        slider_layout.addStretch(1)
-        
-        # Add slider container to responsive layout
-        responsive_layout.addWidget(slider_container, 1)
-        
-        # Add responsive container to main layout
-        main_layout.addWidget(responsive_container)
-        
-        # Set resize behavior to adapt to parent size
-        responsive_layout.setStretch(0, 1)
-        responsive_layout.setStretch(1, 1)
+from ui.components.board_components import ChessSquare, ThinkingIndicator
+from ui.components.history import MoveHistoryWidget
+from ui.components.sidebar import AIControlPanel
+from ui.components.popups import PawnPromotionDialog, GameOverPopup
+from ui.components.animations import AnimatedLabel
+from ui.workers import AIWorker
 
 class ChessBoard(QMainWindow):
     def __init__(self, mode="human_ai", parent_app=None):
@@ -1716,42 +874,3 @@ class ChessBoard(QMainWindow):
         
         # Apply the new sizes
         self.main_splitter.setSizes([board_portion, sidebar_portion])
-
-class ChessApp(QApplication):
-    def __init__(self, args):
-        super().__init__(args)
-        self.setStyle("Fusion")  # Use Fusion style for better cross-platform look
-        
-        # Apply custom palette for better overall aesthetics
-        palette = QPalette()
-        palette.setColor(QPalette.Window, QColor(44, 62, 80))
-        palette.setColor(QPalette.WindowText, QColor(236, 240, 241))
-        palette.setColor(QPalette.Base, QColor(255, 255, 255))
-        palette.setColor(QPalette.AlternateBase, QColor(245, 245, 245))
-        palette.setColor(QPalette.Button, QColor(52, 73, 94))
-        palette.setColor(QPalette.ButtonText, QColor(236, 240, 241))
-        self.setPalette(palette)
-        
-        # Load a standard font that will be available on all systems
-        self.setFont(QFont("Arial", 10))
-        
-        self.chess_window = None
-        self.show_start_screen()
-    
-    def show_start_screen(self):
-        """Show the start screen to select game mode"""
-        # Close any existing chess window
-        if self.chess_window:
-            self.chess_window.close()
-            self.chess_window = None
-            
-        start_screen = StartScreen()
-        if start_screen.exec_() == QDialog.Accepted:
-            mode = start_screen.get_mode()
-            if mode:
-                self.chess_window = ChessBoard(mode, self)
-                self.chess_window.show()
-
-if __name__ == "__main__":
-    app = ChessApp(sys.argv)
-    sys.exit(app.exec_())
